pyproject.toml
[tool.poetry]
name = "projeto kafka"
version = "0.1.0"
description = ""
authors = ["VitorPudo <vitor.silva@shopee.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
install = "^1.3.5"
kafka-python = "^2.0.2"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
